requirements.txt
google-genai>=1.65.0
scrapetube>=2.6.0
pandas>=3.0.0
openpyxl>=3.1.0
fuzzywuzzy>=0.18.0
RapidFuzz>=3.0.0
